refactor: recruit times type setting - activityDate : Date - deadLine, updatedAt, createdAt : LocalDateTime(Instant) - activityStart, activityEnd : BigDecimal

+++ backend/ttabong/src/main/java/com/ttabong/dto/recruit/requestDto/org/UpdateRecruitsRequestDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Date;
 import java.util.List;
@@ -17,8 +18,8 @@ public class UpdateRecruitsRequestDto {
     private Integer recruitId;
     private LocalDateTime deadline;
     private Date activityDate;
-    private double activityStart;
-    private double activityEnd;
+    private BigDecimal activityStart;
+    private BigDecimal activityEnd;
     private Integer maxVolunteer;
     private List<String> images;
     private Integer imageCount;
